--- packages/liferay-theme-tasks/lib/divert/7.0/kickstart_prompt_helpers.js
@@ -1,58 +1,0 @@
-'use strict';
-
-const inquirer = require('inquirer');
-const _ = require('lodash');
-
-const GlobalModulePrompt = require('../../prompts/global_module_prompt');
-const NPMModulePrompt = require('../../prompts/npm_module_prompt');
-const themeUtil = require('../../util');
-
-const choices = [
-	{
-		name: 'Search globally installed npm modules',
-		value: 'global',
-	},
-	{
-		name: 'Search npm registry (published modules)',
-		value: 'npm',
-	},
-	new inquirer.Separator(),
-	{
-		name: 'Classic',
-		value: 'classic',
-	},
-];
-
-function _afterPromptThemeSource(answers, promptInstance) {
-	const config = {
-		themelet: false,
-	};
-
-	const themeSource = answers.themeSource;
-
-	if (themeSource === 'npm') {
-		NPMModulePrompt.prompt(
-			config,
-			_.bind(promptInstance._afterPromptModule, promptInstance)
-		);
-	} else if (themeSource === 'global') {
-		GlobalModulePrompt.prompt(
-			config,
-			_.bind(promptInstance._afterPromptModule, promptInstance)
-		);
-	} else if (themeSource === 'classic') {
-		let classicPath = themeUtil.resolveDependency(
-			'liferay-frontend-theme-classic-web',
-			promptInstance.themeConfig.version
-		);
-
-		promptInstance.done({
-			modulePath: classicPath,
-		});
-	}
-}
-
-module.exports = {
-	choices,
-	_afterPromptThemeSource,
-};
