+++ packages/liferay-theme-tasks/lib/lookup.js
@@ -1,0 +1,25 @@
+const lfrThemeConfig = require('./liferay_theme_config');
+
+const DEFAULT_VERSION = '7.1';
+
+/**
+ * Utility for retrieving version-specific functionality.
+ *
+ * @param {string} key The identifying name for the functionality.
+ * @param {string} [version] The Liferay version; will be inferred if omitted.
+ * @return {?}
+ */
+function lookup(key, version = null) {
+	if (!version) {
+		let config = lfrThemeConfig.getConfig();
+
+		version = config ? config.version : DEFAULT_VERSION;
+	}
+
+	return {
+		'kickstart:choices': require('./lookup/kickstart').choices,
+		'kickstart:afterPromptThemeSource': require('./lookup/kickstart').afterPromptThemeSource,
+	}[key](version);
+}
+
+module.exports = lookup;
